Moving message unit tests into their own categories
