+++ src/twitch/tests/messages/cheers.rs
@@ -1,0 +1,28 @@
+use color_eyre::{Result, eyre::ContextCompat};
+
+use crate::twitch::{models::ReceivedTwitchMessagePayload, tests::messages::FULL_CHEER};
+
+#[test]
+fn test_deserialize_full_cheer() -> Result<()> {
+    let raw_message: serde_json::Value = serde_json::from_str(&FULL_CHEER)?;
+    let message = serde_json::from_str::<ReceivedTwitchMessagePayload>(&FULL_CHEER)?;
+
+    let raw_bits = raw_message
+        .pointer("/event/cheer/bits")
+        .context("Could not find cheer bits")?
+        .as_u64()
+        .context("Cheer could not be converted to u64")?;
+
+    let bits = message
+        .event()
+        .clone()
+        .context("Could not find cheer deserialized event")?
+        .cheer()
+        .clone()
+        .context("Could not find cheer in event")?
+        .bits();
+
+    assert_eq!(raw_bits, bits);
+
+    Ok(())
+}
