+++ src/twitch/tests/messages/mod.rs
@@ -1,23 +1,33 @@
 use std::sync::LazyLock;
 
-#[allow(clippy::module_inception)]
-mod messages;
+mod badges;
+mod cheers;
+mod emotes;
 
-pub static BADGES: LazyLock<&str> = LazyLock::new(|| include_str!("data/badges.json"));
-pub static FULL_CHEER: LazyLock<&str> = LazyLock::new(|| include_str!("data/full_cheer.json"));
-pub static FULL_EMOTE: LazyLock<&str> = LazyLock::new(|| include_str!("data/full_emote.json"));
-pub static FULL_MESSAGE_ONE_WORD: LazyLock<&str> =
-    LazyLock::new(|| include_str!("data/full_message_one_word.json"));
-pub static FULL_MESSAGE: LazyLock<&str> = LazyLock::new(|| include_str!("data/full_message.json"));
-pub static MANY_EMOTES: LazyLock<&str> = LazyLock::new(|| include_str!("data/many_emotes.json"));
-pub static MULTIPLE_EMOTES: LazyLock<&str> =
+// Badges
+static BADGES: LazyLock<&str> = LazyLock::new(|| include_str!("data/badges.json"));
+static NO_BADGES: LazyLock<&str> = LazyLock::new(|| include_str!("data/no_badges.json"));
+static INVALID_BADGES: LazyLock<&str> = LazyLock::new(|| include_str!("data/invalid_badges.json"));
+
+// Cheers
+static FULL_CHEER: LazyLock<&str> = LazyLock::new(|| include_str!("data/full_cheer.json"));
+
+// Emotes
+static FULL_EMOTE: LazyLock<&str> = LazyLock::new(|| include_str!("data/full_emote.json"));
+static MANY_EMOTES: LazyLock<&str> = LazyLock::new(|| include_str!("data/many_emotes.json"));
+static MULTIPLE_EMOTES: LazyLock<&str> =
     LazyLock::new(|| include_str!("data/multiple_emotes.json"));
-pub static NO_BADGES: LazyLock<&str> = LazyLock::new(|| include_str!("data/no_badges.json"));
-pub static PARTIAL_EMOTE_MESSAGE: LazyLock<&str> =
+
+// Messages
+static FULL_MESSAGE_ONE_WORD: LazyLock<&str> =
+    LazyLock::new(|| include_str!("data/full_message_one_word.json"));
+static FULL_MESSAGE: LazyLock<&str> = LazyLock::new(|| include_str!("data/full_message.json"));
+
+// Partials
+static PARTIAL_EMOTE_MESSAGE: LazyLock<&str> =
     LazyLock::new(|| include_str!("data/partial_emote_message.json"));
-pub static PARTIAL_MENTION: LazyLock<&str> =
+static PARTIAL_MENTION: LazyLock<&str> =
     LazyLock::new(|| include_str!("data/partial_mention.json"));
-pub static PARTIAL_MESSAGE_EMOTE_NO_ID: LazyLock<&str> =
+static PARTIAL_MESSAGE_EMOTE_NO_ID: LazyLock<&str> =
     LazyLock::new(|| include_str!("data/partial_message_emote_no_id.json"));
-pub static PARTIAL_REPLY: LazyLock<&str> =
-    LazyLock::new(|| include_str!("data/partial_reply.json"));
+static PARTIAL_REPLY: LazyLock<&str> = LazyLock::new(|| include_str!("data/partial_reply.json"));
