+++ src/twitch/tests/messages/partials.rs
@@ -1,0 +1,44 @@
+// #[test]
+// fn test_deserialize_full_emote() {
+//     assert!(deserialize_message(&FULL_EMOTE).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_full_message_one_word() {
+//     assert!(deserialize_message(&FULL_MESSAGE_ONE_WORD).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_full_message() {
+//     assert!(deserialize_message(&FULL_MESSAGE).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_many_emotes() {
+//     assert!(deserialize_message(&MANY_EMOTES).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_multiple_emotes() {
+//     assert!(deserialize_message(&MULTIPLE_EMOTES).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_partial_emote_message() {
+//     assert!(deserialize_message(&PARTIAL_EMOTE_MESSAGE).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_partial_mention() {
+//     assert!(deserialize_message(&PARTIAL_MENTION).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_partial_message_emote_no_id() {
+//     assert!(deserialize_message(&PARTIAL_MESSAGE_EMOTE_NO_ID).is_ok());
+// }
+
+// #[test]
+// fn test_deserialize_partial_reply() {
+//     assert!(deserialize_message(&PARTIAL_REPLY).is_ok());
+// }
